Isolate application in shadow dom
Refactored application for business logic to be in a testable
class (although there are no tests), and all dependencies to
be via dependency injection. Moved the logic for focusing on
a title into application, had to establishing a communication
channel from Toc to the applicaiton runner.

```
TOC -> Nav -> Skeleton -> ApplicationRunner
```

Just emitted an event in TOC and used composed to ensure it
bubbled through the shadow dom of indivisual web components.